Describe template config sharing across many workspace packages
There's a decision point here with regard to whether we want to introduce "merging" of template configs.

Partitioning
============

A solution that would require no merging of configs requires users to create partitions of their workspace packages. The partition would be created with a config that looks like:

```json
{
  "pnpm": {
    "workspaceTemplateConfigs": {
      // Applies to all packages in the workspace that's a "library"
      "library": {
        "authoring": [
          "@organization/authoring-metadata@0.1.0",
          "@team/authoring-metadata@0.1.0"
        ],
        "catalog": ["@example/frontend-catalog@workspace:"],
        "scripts": ["@example/library-scripts@workspace:"]
      },

      // Other packages may be tagged as an "application".
      "application": {
        "authoring": [
          "@organization/authoring-metadata@0.1.0",
          "@team/authoring-metadata@0.1.0"
        ],
        "catalog": ["@example/frontend-catalog@workspace:"],
        "scripts": ["@example/application-scripts@workspace:"]
      }

    },
  }
}
```

A package would only be able to reference one of the predefined template configs from the root `package.json`.

```json5
// packages/foo/package.json
{
  "name": "@example/foo",
  "pnpm": {
    "templates": {
      "workspaceConfig": "library"
    }
  }
}
```

Tagging
=======

A one-to-many relationship between template configs and packages may be too limiting. The duplication in the previous example is already showing. A many-to-many mechanism would allow less template config duplication.

```json5
// packages.json
{
  "pnpm": {
    "workspaceTemplateConfigs": {
      // All packages in the workspace should use this configuration.
      "default": {
        "authoring": [
          "@organization/authoring-metadata@0.1.0",
          "@team/authoring-metadata@0.1.0"
        ],
        "catalog": ["@example/frontend-catalog@workspace:"],
      },

      // Applies to all packages in the workspace tagged as a "library"
      "library": {
        "scripts": ["@example/library-scripts@workspace:"]
      },

      // Other packages may be tagged as an "application". For this example,
      // applications have a different set of scripts.
      "application": {
        "scripts": ["@example/application-scripts@workspace:"]
      }

    },
  }
}
```

```json5
// packages/foo/package.json
{
  "name": "@example/foo",
  "pnpm": {
    "templates": {
      "workspaceConfigs": ["default", "library"]
    }
  }
}
```

The downside is users would have to reason about how multiple template config blocks merge together. In the example of above, `default` and `library` cleanly merge, but that wouldn't be the case if both blocks specified an `authoring` flavor. Should the `authoring` arrays merge, or should the later configs completely replace all prior ones? This is unfortunate new complexity for users to have to reason about.

A `pnpm template view` command that shows the resulting `package.json` and how each field was determined would help, but users may still have to jump between multiple `package.json` files to reason about properties that were previously simple.

It's worth noting that from an implementation standpoint, merging configs is easy. The complexity to be mindful of is what's introduced to users.

Alternative Tagging Syntax
==========================

Instead of a tag → template mapping, an alternative design would be a template → tag mapping.

```json5
// packages.json
{
  "pnpm": {
    "workspaceTemplates": {
      "@organization/authoring-metadata@0.1.0": { "provides": "authoring", "tags": ["default"] },
      "@team/authoring-metadata@0.1.0": { "provides": "authoring", "tags": ["default"] },
      "@example/frontend-catalog@workspace:": { "provides": "catalog", "tags": ["default"] },

      "@example/library-scripts@workspace:": { "provides": "scripts", "tags": ["library"] },
      "@example/application-scripts@workspace:": { "provides": "scripts", "tags": ["library"] }
    },
  }
}
```

The main benefit is that template names don't need to be written more than once. The downside is that it's harder to tell which template applies to each package since users have to look at every entry and each `tags` field.

My thoughts are:

  - It probably is worthwhile to introduce some sort of template config merging.
  - The tag → template mapping is probably preferable over the reverse. In projects using a few amount of templates, template → tag looks nicer. But projects with many templates would be much harder to reason about.